+++ changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.21.5
+- Reworked player equipment